pyproject.toml
[project]
name = "rattler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-sqlalchemy>=3.1.1",
]
